added tests

+++ src/main/java/core/basesyntax/service/impl/FileWriterServiceImpl.java
@@ -2,15 +2,21 @@ package core.basesyntax.service.impl;
 
 import core.basesyntax.service.FileWriterService;
 import java.io.BufferedWriter;
+import java.io.File;
 import java.io.IOException;
 
 public class FileWriterServiceImpl implements FileWriterService {
     @Override
     public void writeToFile(String content, String filePath) {
-        try (BufferedWriter writer = new BufferedWriter(new java.io.FileWriter(filePath))) {
-            writer.write(content);
-        } catch (IOException e) {
-            throw new RuntimeException("Can't find file by path: " + filePath, e);
+        File file = new File(filePath);
+        if (file.exists()) {
+            try (BufferedWriter writer = new BufferedWriter(new java.io.FileWriter(filePath))) {
+                writer.write(content);
+            } catch (IOException e) {
+                throw new RuntimeException("Error writing to file: " + filePath, e);
+            }
+        } else {
+            throw new RuntimeException("File not found: " + filePath);
         }
     }
 }
